Unifying and Cleaning Webpack Files
