+++ package.json
@@ -4,7 +4,7 @@
   "description": "Pokémon Safari and Pokédex developed with React, Redux and Express.",
   "main": "index.js",
   "scripts": {
-    "build": "webpack-cli --config webpack.dev.config.js --colors",
+    "build": "webpack-cli --config webpack.config.js --colors",
     "build:dev": "webpack-dev-server --open --config ./webpack.dev.config.js",
     "build:local": "webpack --env.NODE_ENV=local --env.production --progress",
     "build:prod": "webpack -p --env.NODE_ENV=production --env.production --progress",
