+++ src/frontend/localization/i18n.js
@@ -4,6 +4,7 @@ import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 import { isServer } from 'Utils/functions';
 import { resources } from './translations';
+require('dotenv').config();
 
 i18n
   .use(Backend)// load translation using xhr -> see /public/locales
@@ -13,7 +14,7 @@ i18n
     resources,
     lng: !isServer ? localStorage.getItem('language') || 'en' : 'en',
     fallbackLng: 'en',
-    debug: true,
+    debug: (process.env.NODE_ENV === 'production') ? false : true,
     interpolation: {
       escapeValue: false, // not needed for react as it escapes by default
     },
