+++ webpack.dev.config.js
@@ -1,69 +1,20 @@
 const path = require('path');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const TerserPlugin = require('terser-webpack-plugin');
-const CompressionPlugin = require('compression-webpack-plugin');
-const ManifestPlugin = require('webpack-manifest-plugin');
-const CopyPlugin = require('copy-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 require('dotenv').config();
 
-const isProd = (process.env.NODE_ENV === 'production');
-
-const plugins = [
-  new MiniCssExtractPlugin({
-    filename: isProd ? 'assets/app-[hash].css' : 'assets/app.css',
-  }),
-];
-
-if (isProd) {
-  plugins.push(
-    new CompressionPlugin({
-      test: /\.js$|\.css$/,
-      filename: '[path].gz',
-    }),
-  );
-
-  plugins.push(
-    new ManifestPlugin(),
-  );
-
-  plugins.push(
-    new CopyPlugin([
-      './src/frontend/public/favicon.ico',
-    ]),
-  );
-}
-
 module.exports = {
-  devtool: isProd ? 'hidden-source-map' : 'cheap-source-map',
-  mode: process.env.NODE_ENV,
-  optimization: {
-    minimizer: isProd ? [new TerserPlugin()] : [],
-    splitChunks: {
-      chunks: 'async',
-      name: true,
-      cacheGroups: {
-        vendors: {
-          name: 'vendors',
-          chunks: 'all',
-          reuseExistingChunk: true,
-          priority: 1,
-          filename: isProd ? 'assets/vendors-[hash].js' : 'assets/vendors.js',
-          enforce: true,
-          test(module, chunks) {
-            const name = module.nameForCondition && module.nameForCondition();
-            return chunks.some((chunk) => chunk.name !== 'vendor' && /[\\/]node_modules[\\/]((?!(react-flags-select)).*)[\\/]((?!(intersection-observer)).*)[\\/]/.test(name));
-          },
-        },
-      },
+  resolve: {
+    alias: {
+      Utils: path.resolve(__dirname, 'src/frontend/utils/'),
+      Constants: path.resolve(__dirname, 'src/frontend/constants/'),
     },
   },
   entry: {
     app: path.resolve(__dirname, 'src/frontend/index.js'),
   },
   output: {
-    path: isProd ? path.join(process.cwd(), './src/server/public') : '/', //path.resolve(__dirname, 'dist'),
-    filename: isProd ? 'assets/app-[hash].js' : 'assets/app.js', //'js/[name].js',
-    publicPath: '/',
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'js/[name].js',
   },
   devServer: {
     port: process.env.PORT || 3001,
@@ -93,22 +44,16 @@ module.exports = {
       },
       {
         test: /\.css$/,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-          },
-          'css-loader',
-        ],
+        use: ['style-loader', 'css-loader'],
       },
       {
         test: /\.(jpg|jpeg|png|gif)$/,
         use: {
-          loader: 'file-loader',
+          loader: 'url-loader',
           options: {
-            // limit: 1000000,
-            // fallback: 'file-loader',
-            name: 'assets/images/[name].[ext]',
-            // name: 'assets/images/[name].[hash].[ext]',
+            limit: 1000000,
+            fallback: 'file-loader',
+            name: 'images/[name].[hash].[ext]',
           },
         },
       },
@@ -116,13 +61,14 @@ module.exports = {
         test: /\.svg/,
         use: {
           loader: 'svg-url-loader',
-          options: {
-            limit: 1000,
-            name: 'assets/images/[name].[ext]',
-          },
         },
       },
     ],
   },
-  plugins,
+  plugins: [
+    new HtmlWebpackPlugin({
+      filename: './index.html',
+      template: './src/frontend/public/index.html',
+    }),
+  ],
 };
